Remove stupid autoclose feature
Ok, so story time.

Back when I discovered that the database closed at 4am and caused errors
with anything running at the time, I was kinda not up to speed on the
exacts. Therefore, because this thing had nobody around to turn it off /
on (apparently it runs overnight?) I got scared of being sent 500000
error emails overnight. Thus, I put a prevention feature in place. After
5 errors, it closes.

Now, I know this is a bad idea, but back then, the only error I
considered after numerous tests was the after 4am error. This never came
to pass. EVER. We just... haven't had issues with it. But we ARE having
issues with the worker, which autoclosed it as a result. Time to remove
this. It's been a long time coming...